requirements.txt
# Core dependencies
dspy-ai>=2.4.0
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Databricks (optional - for production)
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0

# ML tracking
mlflow>=2.10.0

# LLM providers
openai>=1.10.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
